refactor(experiments): extract getMetricTag into a module.
