+++ frontend/src/scenes/experiments/MetricsView/DeltaChart.test.tsx
@@ -1,14 +1,6 @@
-import {
-    ExperimentFunnelsQuery,
-    ExperimentMetric,
-    ExperimentMetricType,
-    ExperimentTrendsQuery,
-    NodeKind,
-} from '~/queries/schema/schema-general'
-import { ExperimentMetricMathType } from '~/types'
+import { ExperimentMetric, ExperimentMetricType, NodeKind } from '~/queries/schema/schema-general'
 
-import { getDefaultMetricTitle, getMetricTag } from './DeltaChart'
-import { generateViolinPath } from './DeltaChart'
+import { generateViolinPath, getDefaultMetricTitle } from './DeltaChart'
 
 /**
  * Helper function to parse SVG path string into points, optionally skipping the initial move command
@@ -186,39 +178,5 @@ describe('getDefaultMetricTitle', () => {
         }
 
         expect(getDefaultMetricTitle(metric)).toBe('purchase')
-    })
-})
-
-describe('getMetricTag', () => {
-    it('handles different metric types correctly', () => {
-        const experimentMetric: ExperimentMetric = {
-            kind: NodeKind.ExperimentMetric,
-            metric_type: ExperimentMetricType.MEAN,
-            metric_config: {
-                kind: NodeKind.ExperimentEventMetricConfig,
-                math: ExperimentMetricMathType.TotalCount,
-                event: 'purchase',
-            },
-        }
-
-        const funnelMetric: ExperimentFunnelsQuery = {
-            kind: NodeKind.ExperimentFunnelsQuery,
-            funnels_query: {
-                kind: NodeKind.FunnelsQuery,
-                series: [],
-            },
-        }
-
-        const trendMetric: ExperimentTrendsQuery = {
-            kind: NodeKind.ExperimentTrendsQuery,
-            count_query: {
-                kind: NodeKind.TrendsQuery,
-                series: [],
-            },
-        }
-
-        expect(getMetricTag(experimentMetric)).toBe('Mean')
-        expect(getMetricTag(funnelMetric)).toBe('Funnel')
-        expect(getMetricTag(trendMetric)).toBe('Trend')
     })
 })
